Moved more code
