+++ bases/renku_data_services/data_api/app.py
@@ -38,7 +38,7 @@ from renku_data_services.search.blueprints import SearchBP
 from renku_data_services.search.reprovision import SearchReprovision
 from renku_data_services.search.solr_user_query import UsernameResolve
 from renku_data_services.session.blueprints import BuildsBP, EnvironmentsBP, SessionLaunchersBP
-from renku_data_services.storage.blueprints import StorageSchemaBP
+from renku_data_services.storage.blueprints import ProjectStorageBP, StorageSchemaBP
 from renku_data_services.users.blueprints import KCUsersBP, UserPreferencesBP, UserSecretsBP
 
 
@@ -222,6 +222,7 @@ def register_all_handlers(app: Sanic, dm: DependencyManager) -> Sanic:
         oauth_client_factory=dm.oauth_http_client_factory,
         project_repo=dm.project_repo,
         project_session_secret_repo=dm.project_session_secret_repo,
+        project_storage_repo=dm.project_storage_repo,
         rp_repo=dm.rp_repo,
         session_repo=dm.session_repo,
         user_repo=dm.kc_user_repo,
@@ -280,7 +281,6 @@ def register_all_handlers(app: Sanic, dm: DependencyManager) -> Sanic:
         data_service_base_url=dm.config.nb_config.data_service_url,
         k8s_client=dm.k8s_client,
         deposit_config=dm.config.deposit_config,
-        project_storage_k8s=dm.project_storage_k8s,
     )
     notifications = NotificationsBP(
         name="notifications",
@@ -310,6 +310,13 @@ def register_all_handlers(app: Sanic, dm: DependencyManager) -> Sanic:
         internal_authenticator=dm.internal_authenticator,
         internal_token_mint=dm.internal_token_mint,
         internal_scope_verifier=dm.internal_scope_verifier,
+    )
+    project_storage = ProjectStorageBP(
+        name="project_storage",
+        url_prefix=url_prefix,
+        project_storage_k8s=dm.project_storage_k8s,
+        project_storage_repo=dm.project_storage_repo,
+        authenticator=dm.authenticator,
     )
     app.blueprint(
         [
@@ -342,6 +349,7 @@ def register_all_handlers(app: Sanic, dm: DependencyManager) -> Sanic:
             capacity_reservation.blueprint(),
             resource_usage.blueprint(),
             internal_authentication.blueprint(),
+            project_storage.blueprint(),
         ]
     )
     if builds is not None:
