+++ test/bases/renku_data_services/data_api/test_data_connectors.py
@@ -2665,685 +2665,6 @@ async def link_dc_project(
     return response
 
 
-@pytest.mark.asyncio
-async def test_post_storage_success(
-    sanic_client: SanicASGITestClient,
-    create_project,
-    admin_headers: dict[str, str],
-    user_headers: dict[str, str],
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-
-    assert response.status_code == 201, response.text
-    assert response.json is not None
-    storage = response.json
-    assert storage.get("project_id") == project["id"]
-    assert storage.get("size") == 10
-    assert storage.get("mount_path") == "/data"
-    assert storage.get("created_by") == "user"
-    assert "ETag" in response.headers
-
-
-@pytest.mark.asyncio
-async def test_post_storage_unauthenticated_fails(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", json=payload)
-
-    assert response.status_code == 401, response.text
-
-
-@pytest.mark.asyncio
-async def test_post_storage_not_allowed_fails(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", json=payload, headers=user_headers)
-
-    assert response.status_code == 403, response.text
-
-
-@pytest.mark.asyncio
-async def test_post_storage_duplicate_fails(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status == 201
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-
-    assert response.status_code == 422, response.text
-
-
-@pytest.mark.asyncio
-async def test_get_one_storage_success(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/{storage_id}", headers=user_headers)
-
-    assert response.status_code == 200, response.text
-    assert response.json is not None
-    storage = response.json
-    assert storage.get("id") == storage_id
-    assert storage.get("project_id") == project["id"]
-    assert storage.get("size") == 10
-    assert storage.get("mount_path") == "/data"
-
-
-@pytest.mark.asyncio
-async def test_get_one_storage_not_found(sanic_client: SanicASGITestClient, user_headers: dict[str, str]) -> None:
-    from ulid import ULID
-
-    non_existent_id = str(ULID())
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/{non_existent_id}", headers=user_headers)
-
-    assert response.status_code == 404, response.text
-
-
-@pytest.mark.asyncio
-async def test_get_one_storage_etag(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-    etag = response.headers["ETag"]
-
-    headers = merge_headers(user_headers, {"If-None-Match": etag})
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/{storage_id}", headers=headers)
-
-    assert response.status_code == 304, response.text
-
-
-@pytest.mark.asyncio
-async def test_get_storage_to_project_for_no_storage(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-
-    _, response = await sanic_client.get(f"/api/data/projects/{project['id']}/storage", headers=user_headers)
-
-    assert response.status_code == 200, response.text
-    assert response.json == []
-
-
-@pytest.mark.asyncio
-async def test_get_storage_to_project_success(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status == 201
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.get(f"/api/data/projects/{project['id']}/storage", headers=user_headers)
-
-    assert response.status_code == 200, response.text
-    assert len(response.json) == 1
-    storage = response.json[0]
-    assert storage.get("project_id") == project["id"]
-    assert storage.get("size") == 10
-    assert storage.get("mount_path") == "/data"
-
-
-@pytest.mark.asyncio
-async def test_delete_storage_success(
-    sanic_client: SanicASGITestClient,
-    create_project,
-    user_headers: dict[str, str],
-    admin_headers: dict[str, str],
-    cluster,
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    _, response = await sanic_client.delete(f"/api/data/data_connectors/storage/{storage_id}", headers=user_headers)
-    assert response.status_code == 204, response.text
-
-    _, response = await sanic_client.get(f"/api/data/projects/{project['id']}/storage", headers=user_headers)
-    assert response.status_code == 200, response.text
-    assert response.json == []
-
-
-@pytest.mark.asyncio
-async def test_delete_storage_unauthenticated_fails(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    _, response = await sanic_client.delete(f"/api/data/data_connectors/storage/{storage_id}")
-
-    assert response.status_code == 401, response.text
-
-
-@pytest.mark.asyncio
-async def test_post_storage_allow_success(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-
-    assert response.status_code == 201, response.text
-    assert response.json is not None
-    allow = response.json
-    assert allow.get("project_id") == project_id
-    assert allow.get("max_size") == 10
-
-
-@pytest.mark.asyncio
-async def test_post_storage_allow_requires_admin(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-
-    payload = {"project_id": project["id"], "max_size": 10}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage/allow", headers=user_headers, json=payload)
-
-    assert response.status_code == 403, response.text
-
-
-@pytest.mark.asyncio
-async def test_post_storage_allow_unauthenticated_fails(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-
-    payload = {"project_id": project["id"], "max_size": 10}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage/allow", json=payload)
-
-    assert response.status_code == 401, response.text
-
-
-@pytest.mark.asyncio
-async def test_post_storage_allow_duplicate_fails(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-
-    assert response.status_code == 422, response.text
-
-
-@pytest.mark.asyncio
-async def test_delete_storage_allow_success(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.delete(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=admin_headers
-    )
-
-    assert response.status_code == 204, response.text
-
-    # Re-adding after deletion should succeed
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-
-@pytest.mark.asyncio
-async def test_delete_storage_allow_requires_admin(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.delete(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=user_headers
-    )
-
-    assert response.status_code == 403, response.text
-
-
-@pytest.mark.asyncio
-async def test_get_storage_allow_success(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}", headers=user_headers)
-
-    assert response.status_code == 200, response.text
-    assert response.json is not None
-    assert response.json.get("project_id") == project_id
-    assert response.json.get("max_size") == 10
-
-
-@pytest.mark.asyncio
-async def test_get_storage_allow_not_in_list(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}", headers=user_headers)
-
-    assert response.status_code == 404, response.text
-
-
-@pytest.mark.asyncio
-async def test_get_storage_allow_unauthenticated(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}")
-
-    assert response.status_code == 401, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_success(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 5, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage = response.json
-    storage_id = storage["id"]
-    original_etag = response.headers["ETag"]
-
-    # Patch the size
-    headers = merge_headers(user_headers, {"If-Match": original_etag})
-    patch = {"size": 8}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 200, response.text
-    assert response.json is not None
-    updated_storage = response.json
-    assert updated_storage.get("id") == storage_id
-    assert updated_storage.get("size") == 8
-    assert updated_storage.get("mount_path") == "/data"
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_mount_path(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage = response.json
-    storage_id = storage["id"]
-
-    # Patch the mount path
-    headers = merge_headers(user_headers, {"If-Match": response.headers["ETag"]})
-    patch = {"mount_path": "/new/mount"}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 200, response.text
-    assert response.json is not None
-    updated_storage = response.json
-    assert updated_storage.get("id") == storage_id
-    assert updated_storage.get("size") == 10
-    assert updated_storage.get("mount_path") == "/new/mount"
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_both_fields(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 5, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage = response.json
-    storage_id = storage["id"]
-
-    # Patch both fields at once
-    headers = merge_headers(user_headers, {"If-Match": response.headers["ETag"]})
-    patch = {"size": 8, "mount_path": "/new/mount"}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 200, response.text
-    assert response.json is not None
-    updated_storage = response.json
-    assert updated_storage.get("id") == storage_id
-    assert updated_storage.get("size") == 8
-    assert updated_storage.get("mount_path") == "/new/mount"
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_without_if_match_header(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    # Patch without If-Match header
-    patch = {"size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=user_headers, json=patch
-    )
-
-    assert response.status_code == 428, response.text
-    assert "If-Match header not provided" in response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_with_invalid_etag(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 5, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-    correct_etag = response.headers["ETag"]
-
-    # Patch with wrong ETag
-    headers = merge_headers(user_headers, {"If-Match": "wrong-etag"})
-    patch = {"size": 8}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 409, response.text
-
-    # Verify the etag changed after a successful patch
-    headers = merge_headers(user_headers, {"If-Match": correct_etag})
-    patch = {"size": 6}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-    assert response.status_code == 200, response.text
-    new_etag = response.headers["ETag"]
-    assert new_etag != correct_etag
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_not_found(sanic_client: SanicASGITestClient, user_headers: dict[str, str]) -> None:
-    from ulid import ULID
-
-    non_existent_id = str(ULID())
-    headers = merge_headers(user_headers, {"If-Match": "some-etag"})
-    patch = {"size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{non_existent_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 404, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_unauthenticated_fails(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 5, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    # Patch without authentication
-    _, response = await sanic_client.patch(f"/api/data/data_connectors/storage/{storage_id}", json={"size": 8})
-
-    assert response.status_code == 401, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_exceeds_max_size(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 5, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    # Try to patch size beyond the allowed max (10GB)
-    headers = merge_headers(user_headers, {"If-Match": response.headers["ETag"]})
-    patch = {"size": 11}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 422, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_invalid_mount_path(
-    sanic_client: SanicASGITestClient, create_project, user_headers: dict[str, str], admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    namespace = f"{project['namespace']}/{project['slug']}"
-
-    project_id = project["id"]
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    payload = {"namespace": namespace, "size": 10, "mount_path": "/data"}
-    _, response = await sanic_client.post("/api/data/data_connectors/storage", headers=user_headers, json=payload)
-    assert response.status_code == 201, response.text
-    storage_id = response.json["id"]
-
-    # Try to patch with invalid mount path
-    headers = merge_headers(user_headers, {"If-Match": response.headers["ETag"]})
-    patch = {"mount_path": "/etc/passwd"}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/{storage_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 422, response.text
-
-
 async def test_get_all_dc_links(
     sanic_client: SanicASGITestClient,
     regular_user: UserInfo,
@@ -3387,215 +2708,3 @@ async def test_get_all_dc_links(
     assert res.status_code == 200, res.text
     assert len(res.json) == 2
     assert p2["id"] not in [i["id"] for i in res.json]
-
-
-# --- PATCH Project Storage Allow Tests ---
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_success(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    # Get the allow entry to retrieve the etag (use user_headers as admin may not have read access)
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}", headers=user_headers)
-    assert response.status_code == 200, response.text
-    etag = response.headers["ETag"]
-
-    # Patch the max_size
-    headers = merge_headers(admin_headers, {"If-Match": etag})
-    patch = {"max_size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 200, response.text
-    assert response.json is not None
-    updated_allow = response.json
-    assert updated_allow.get("project_id") == project_id
-    assert updated_allow.get("max_size") == 20
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_requires_admin(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    # Get the allow entry to retrieve the etag (use user_headers as admin may not have read access)
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}", headers=admin_headers)
-    assert response.status_code == 200, response.text
-    etag = response.headers["ETag"]
-
-    # Try to patch as non-admin
-    headers = merge_headers(user_headers, {"If-Match": etag})
-    patch = {"max_size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 403, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_without_if_match_header(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    # Patch without If-Match header
-    patch = {"max_size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=admin_headers, json=patch
-    )
-
-    assert response.status_code == 428, response.text
-    assert "If-Match header not provided" in response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_with_invalid_etag(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    # Get the allow entry to retrieve the etag (use user_headers as admin may not have read access)
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}", headers=user_headers)
-    assert response.status_code == 200, response.text
-    correct_etag = response.headers["ETag"]
-
-    # Patch with wrong ETag
-    headers = merge_headers(admin_headers, {"If-Match": "wrong-etag"})
-    patch = {"max_size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 409, response.text
-
-    # Verify the etag changed after a successful patch
-    headers = merge_headers(admin_headers, {"If-Match": correct_etag})
-    patch = {"max_size": 15}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=headers, json=patch
-    )
-    assert response.status_code == 200, response.text
-    new_etag = response.headers["ETag"]
-    assert new_etag != correct_etag
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_not_in_list(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    headers = merge_headers(admin_headers, {"If-Match": "some-etag"})
-    patch = {"max_size": 20}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 404, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_unauthenticated_fails(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    # Patch without authentication
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", json={"max_size": 20}
-    )
-
-    assert response.status_code == 401, response.text
-
-
-@pytest.mark.asyncio
-async def test_patch_storage_allow_min_size(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    # Get the allow entry to retrieve the etag (use user_headers as admin may not have read access)
-    _, response = await sanic_client.get(f"/api/data/data_connectors/storage/allow/{project_id}", headers=user_headers)
-    assert response.status_code == 200, response.text
-    etag = response.headers["ETag"]
-
-    # Try to set max_size below minimum (1GB)
-    headers = merge_headers(admin_headers, {"If-Match": etag})
-    patch = {"max_size": 0}
-    _, response = await sanic_client.patch(
-        f"/api/data/data_connectors/storage/allow/{project_id}", headers=headers, json=patch
-    )
-
-    assert response.status_code == 422, response.text
-    assert "at least 1GB" in response.text
-
-
-@pytest.mark.asyncio
-async def test_get_all_storage_allow(
-    sanic_client: SanicASGITestClient, create_project, admin_headers: dict[str, str], user_headers: dict[str, str]
-) -> None:
-    project = await create_project(sanic_client, "Test Project")
-    project_id = project["id"]
-
-    payload = {"project_id": project_id, "max_size": 10}
-    _, response = await sanic_client.post(
-        "/api/data/data_connectors/storage/allow", headers=admin_headers, json=payload
-    )
-    assert response.status_code == 201, response.text
-
-    _, response = await sanic_client.get("/api/data/data_connectors/storage/allow", headers=admin_headers)
-    assert response.status_code == 200, response.text
-    assert response.json is not None, f"No json response body: {response.text}"
-    assert isinstance(response.json, list)
-    assert len(response.json) == 1
-    assert response.json[0]["project_id"] == project_id
-    assert response.json[0]["max_size"] == 10
